Moved old main window classes into legacy folder.
