+++ legacy/src/View/mainpage/ClinicalDataDisplay.py
