+++ legacy/src/View/mainpage/ClinicalDataForm.py
