+++ legacy/src/View/mainpage/PatientBar.py
