+++ legacy/src/View/mainpage/StructureTab.py
